backend/requirements.txt
asgiref==3.8.1
boto3==1.34.94
botocore==1.34.94
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
gunicorn==21.2.0
heyoo==0.1.2
idna==3.6
jmespath==1.0.1
packaging==24.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
requests==2.31.0
requests-toolbelt==1.0.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
imap_tools
